Added iam roles to cloud profiles
Updated cloud profiles to include an IAM profile so that if needed in
the future then nodes don't need to be relaunched. This is because the
only time that and IAM profile can be assigned to an instance is at creation.

Updated orchestration to be able to build cloud instances, as well as
making sure that an IAM profile is present for each instance type.

Added state to install libgit2 and pygit2 from source to use latest version.