Forcing virtualenv to depend on install of python
In order to ensure that the virtualenv for ansible is created with an
updated version of Python I added a require statement pointing to the
installation of system packages. This is to work around an issue with
the venv being created with Python v2.7.3 which fails on a second run
due to the system Python being upgraded to 2.7.10 by Ansible.