Moved AWS VPC creation to separate state
Moved the creation of the Micromasters VPC in AWS in order to allow for
the orchestration run to complete. It was failing due to the fact that
it was trying to populate the list of subnet ids before the subnets were
created, but that list is necessary for creation of the map template to
provision the EC2 instances.