Added symlink back in to set proper directory
The edxapp playbook attempts to create the
`/edx/var/edxapp/course_static` directory but fails when the symlink
is in place. The ansible var that controls that directory creation
isn't used anywhere else, so it is safe to make it create an unused
directory as long as we create the symlink with the path that edX is
expecting to be present.