Fixing consul agent timeout errors
The consul agents need to have port 8301 open in order for the gossip
protocol to operate properly. Because it was not open it causes
occasional dropouts leading to erroneous error alerts in DataDog. This
commit adds a new security group to expose port 8300 and 8301 to all
nodes running the consul agent.