Updated the edX orchestrate states to DRY them up
Extracted a Jinja template for setting the VPC attributes to simplify
writing of reusable orchestrate states. Consolidated the orchestrate
states to work on multiple VPCs by passing a `VPC_NAME` environment
variable when calling the orchestrate run.

This declutters the repository as well as making it more explicit when
running orchestrate states against the production environment since
the default VPC is MITx QA.