Updated to set file type and add dummy snapshot id for root volumes
When launching new instances with the Debian 9 AMI they were
complaining about duplicate device names. Updated the snapshot ID in
the profiles to point to a dummy snapshot as a means of avoiding that
issue.

Also updated the file metadata to properly identify the fact that the
profiles are YAML