Added state to control pre-commit hook template for git exports
The git exports from edX for courses in the split-mongo modulestore will no longer include the term as part of the exported information. To allow for us to still use the course content in our LMS environment we need to run a pre-commit hook that places that information back in the courseware. This state adds a global pre-commit hook template which gets included in the exported git repository so that it is executed as part of our existing workflow.