Migrating pillar data to our public salt repository
Now that we have removed all of the encrypted secrets from our configuration data in favor of calls to Vault we can release it in our public repository. Co-locating our pillar data with the rest of our salt code simplifies tasks and code reviews that require modifying the configuration data along with the state logic.