Moved consul states to top of top file to fix ordering errors
The Consul daemon needs to be running before a majority of systems can properly function due to needing to be able to reach external services at launch. Moved the consul states to the top of the top file to ensure that Consul gets deployed first to avoid first run failures during highstates.