Removed techtv data source as it frequently fails to render
The MySQL host frequently throws an error about having too many open connections, requiring the `flush-hosts` command to be executed and making the pillar rendering fail. Since this source is not frequently used and is likely to be retired soon I removed it from the pillar data, in favor of manually updating the credentials as needed.