Moved sandbox version values
Pillar is failing to render with `maximum recursion depth exceeded in cmp` pointing to a line that is trying to resolve the value for the theme key. Moved those values right under versions just to try out whether this resolves it (don't have a lot of a confidence that ti will).