Added workaround for CAS in production WRT entityID and domain names
The CAS metadata is registered with the entityID for auth.mitx.mit.edu, but uses cas.mitx.mit.edu for the domain in the production configuration for residential. The shibboleth pillar is defaulting to using the first defined domain in the domains list, so I set that to be the auth subdomain, however when using both domains in the Nginx config it was causing a redirect to auth.mitx.mit.edu after completing the Touchstone login, so I overrode the Nginx config to only use the second defined domain in production.