Added way to customize Consul service name for RDS
In order to allow for mapping the new pattern for building RDS instances to the old Consul service name I added a way to explicitly set the service name as part of the db configuration. This allows for preserving the Ansible configuration of edX while still using current deployment tooling.