Updated settings for building bucket names to work with xpro
The settings and names that we were using for S3 in the edX configs had a lot of assumptions about the residential environment built into them. I updated the settings to be a bit more granular in how they get composed so that we can more easily identify which buckets go with which environments and application stacks.