Updated settings and hacks to resolve SAML login
In order for us to be able to create new accounts for users using SAML they need to be able to self-register as part of the auth flow. This requires us to remove our forced redirect of the `/register` endpoint and set the feature flag to allow account creation. This will likely require further efforts to prevent creation of accounts by any means other than SAML.