Make adjustments for updated mirror scripts
Make adjustments to pillars and states to account for recent updates in
the `ocwcms` repository to files that had been out-of-sync between the
mirror server and repository.

Add some /etc/hosts entries for hardcoded hostnames in scripts that are
supposed to run on the mirror server, where there is no engines.conf
file to parameterize such strings.

See `ocwcms` commits:
https://github.com/mitocw/ocwcms/commit/435b26a4daffb87845a88cc687968b7298f874c0
https://github.com/mitocw/ocwcms/commit/d98a4919813cd97103dce3ee1a7a41ae359eac15