Fix OCW origin server's fluentD access log parsing
Fix the parsing attributes for the fluentD configuration that parses OCW
origin server Nginx access logs.

Use the fluentD bundled 'apache2' type, which is a misnomer: it's the
standard Combined Log Format, which is used by our Nginx servers. It's
the same log format that has been working fine with `ocs_cms.sls`.