Fix fluentD parsing of Nginx access log null value
Fix the regex that fluentD uses to determine a null value in a field in
an Nginx access log.

It takes a field with a whole value of "-" and stores it as null. The
problem was that the null_value_pattern parameter is a regex, so it was
registering anything _containing_ a hyphen as null.