Fix FluentD shipping, OCW Zope and Apache access logs
Fix how FluentD ships Apache2-format access logs from OCW's Zope and
Apache acccess logs.

Stop it from sending the `time' field, which has special significance
for Elasticsearch, which tries to parse it as a long int and fails to
save the log message with a parsing error.

FluentD's default for the `keep_time_key' parameter is `false', but we
had it set to `true'.