Updated app deployments to use release ids
Added a release-id specifier in instance/minion names in order to simplify build/deploy of instances and applications.

As things stand currently, if we wish to deploy a new set of instances it requires either increasing the instance count temporarily, or performing some manual intervention. By adding an updateable and arbitrary release-id it makes it easier to deploy multiple releases in parallel.