Adding setup for bootstrapping a master server from scratch
In order to be able to deploy salt masters from scratch we need to have a mechanism that isn't depending on a running master. This is the initial steps for adding a directory structure that allows for running from a local machine as well as splitting the pillar data such that it won't fail to render before being joined to a Vault server.