Updated Consul DNS for operations services to simplify discovery
With the introduction of the operations-qa environment there were a number of hardcoded DNS records that would fail to resolve for QA services. Added a query template to allow for querying operations and operations-qa environments based on what is reachable to the given client, removing the need to generate a mapping of which envs talk to which other envs.