Fix OCW Mirror Nginx pillar rendering
Fix the Nginx pillar for OCW's mirror servers. Though there is a "qa" minion
for the ocw-mirror role, its Nginx configuration is superficial, because the
only reason for the "qa" mirror to exist is to verify that mirror publishing
scripts work. The QA mirror isn't accessed over HTTP. Therefore, there only
has to be one "production" Nginx configuration that actually works.

Without the modification in this commit, there is a Jinja2 rendering
error about a dict having no 'qa' attribute. That's because the
ocw-environment grain on the QA minion is "qa" and there's no such
property of environments:ocw:purposes:ocw-mirror.