Try to make edX Ansible work on Python 3
* Install python3-apt Ubuntu package
* Specify ansible_python_interpreter Ansible variable to try to get
  ansible-playbook to run on the virtualenv's Python.

ansible-playbook has apparently been running on Python 2 in
next-residential with Juniper Alpha, whereas we want it to run on the
virtualenv's Python 3.

When it's made to do so by running it manually on the server, it doesn't
have the `apt' and `apt_pkg' Python modules that `python3-apt' is
supposed to provide.