Increase Caddy timeout
Increase the execution timeout for the webhook script that's run by
Caddy, at least temporarily, until we can figure out what's wrong with
the expression that is supposed to filter on {json.ref}. Because we're
running for every push to the repository (due to the fact that we can't
filter on the branch) the webhook script keeps rerunning, which appears
to result in Caddy eventually killing it in the middle of the 2nd or 3rd
run, after an hour has passed. When that happens, a stale lock file and
"retry" file are left behind.

When subsequent runs are triggered, all they do is touch the retry file,
instead of executing, which means that deployments stop happening. Maybe
the script should have more thorough logic, but I assume that this
problem will go away if we can fix the issue with the expression.