Fix OCW build script exec command
In the OCW build script, fix the `exec' command that re-runs the script.
It is invoked when the script has already cd'd into the
ocw-course-hugo-starter directory, and $0 is relative to that directory,
so `exec $0` tries to find webhook-publish.sh in the wrong place. Let's
just hardcode it, because we already rely on elements of this
architecture being in predictable places and this is fair in our
use-case.