Fix "merge" calls in Vector VRL for MITx
Fix the error-handling for a couple of merge() calls. It doesn't make
any sense to me, but though the current documentation says that this
function is "infallible" (meaning that it does not raise an exception),
one of the function calls is failed by `vector validate` if it does not
have the "abort" syntax for raising exceptions. I'm not sure why one
instance of the function call requires it and the other fails if it has
it; but this is how it passes validation at the moment.