Updating edxapp and CAS database hosts to use direct endpoint
While we are migrating our Consul clusters to the new deployment method we need to account for database endpoints that aren't yet being managed by Pulumi and populated in the service catalog. This moves to looking up the endpoint for the RDS instances used by edxapp and the CAS app so that we don't lose connectivity when rebuilding the Consul cluster.