Adding mapping of Consul datacenter names to Pulumi
As we are migrating things to Pulumi one of the points of mismatch is the datacenter name that we want to use between Salt and Pulumi. Specifically we run into this for xpro and apps environments. This adds specific overrides to allow existing infrastructure to continue working properly as we progress through the migration.