Updated xqueue-watcher configs to map to live/staging split
For our deployment of the Maple release we are splitting the live and staging
environments into separate VPCs. This means that for residential use of xqueue-watcher
we need to deploy a separate group of watchers for those courses that use the staging
environment. This updates the configuration to reduce the amount of extraneous logic so
that we can deploy dedicated watchers per environment.

- Gets rid of iteration over purposes that is difficult to understand
- Adds new environment definitions to include the staging environments of QA and
- Production
- Updates xqueue-watcher versions to match the latest code