Remove RC databases from Redash for MM and MITx Online
The RC databases moved to be on RDS instances. Because of the way that the network setup
works and the Vault credentials management it is not straightforward to use Salt for
managing those connections. This removes that setup from Salt so that the proper configs
don't get overwritten when updating credentials for other data sources.